requirements.txt
channels==2.1.6
autobahn==0.18.1
channels-redis==2.3.2
asgiref==2.3.2
daphne==2.2.4
chatterbot==0.8.7
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-braces==1.13.0
django-cors-headers==2.3.0
django-filter==1.1.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-user-agents==0.3.2
django-webpack-loader==0.5.0
djangorestframework==3.7.7
idna==2.7
Markdown==2.6.10
oauthlib==2.1.0
psycopg2==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
rest-social-auth==1.3.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
ua-parser==0.8.0
urllib3==1.23
user-agents==1.1.0
